+++ sydtest-sqitch-postgres/default.nix
@@ -1,7 +1,7 @@
 { mkDerivation, base, bytestring, containers, lib, monad-logger
 , network-uri, path, path-io, persistent, postgres-options, sydtest
 , sydtest-discover, sydtest-persistent-postgresql, text
-, tmp-postgres, typed-process, unliftio
+, typed-process, unliftio
 }:
 mkDerivation {
   pname = "sydtest-sqitch-postgres";
@@ -10,11 +10,11 @@ mkDerivation {
   libraryHaskellDepends = [
     base bytestring containers monad-logger network-uri path persistent
     postgres-options sydtest sydtest-persistent-postgresql text
-    tmp-postgres typed-process
+    typed-process
   ];
   testHaskellDepends = [
-    base containers path path-io persistent sydtest
-    sydtest-persistent-postgresql text unliftio
+    base path path-io sydtest sydtest-persistent-postgresql text
+    unliftio
   ];
   testToolDepends = [ sydtest-discover ];
   homepage = "https://github.com/NorfairKing/sydtest#readme";